+++ src/main/java/org/patternfly/component/button/Button.java
@@ -67,13 +67,12 @@ import static org.patternfly.layout.Size.lg;
 import static org.patternfly.layout.Size.md;
 
 /**
- * A button is a box area or text that communicates and triggers user actions when clicked or selected. Buttons can be
- * used to communicate and immediately trigger actions a user can take in an application, like submitting a form,
- * canceling a process, or creating a new object. Buttons can also be used to take a user to a new location, like
- * another page inside a web application, or an external site such as help or documentation.
+ * A button is a box area or text that communicates and triggers user actions when clicked or selected. Buttons can be used to
+ * communicate and immediately trigger actions a user can take in an application, like submitting a form, canceling a process,
+ * or creating a new object. Buttons can also be used to take a user to a new location, like another page inside a web
+ * application, or an external site such as help or documentation.
  *
- * @see <a href=
- * "https://www.patternfly.org/components/button/html">https://www.patternfly.org/components/button/html</a>
+ * @see <a href= "https://www.patternfly.org/components/button/html">https://www.patternfly.org/components/button/html</a>
  */
 public class Button extends BaseComponent<HTMLElement, Button>
         implements Disabled<HTMLElement, Button>, Inline<HTMLElement, Button>, Plain<HTMLElement, Button>,
@@ -277,9 +276,8 @@ public class Button extends BaseComponent<HTMLElement, Button>
     // ------------------------------------------------------ public API
 
     /**
-     * Changes the text of this button. Prefer this method over
-     * {@link org.jboss.elemento.HasElement#textContent(String)}, since this method doesn't remove a possible progress
-     * spinner.
+     * Changes the text of this button. Prefer this method over {@link org.jboss.elemento.HasElement#textContent(String)}, since
+     * this method doesn't remove a possible progress spinner.
      */
     public Button text(String text) {
         // just using textContent(text) would remove a possible progress spinner