+++ src/main/java/org/patternfly/component/code/CodeEditorLink.java
@@ -1,0 +1,84 @@
+/*
+ *  Copyright 2023 Red Hat
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package org.patternfly.component.code;
+
+import org.patternfly.component.ComponentReference;
+import org.patternfly.component.SubComponent;
+import org.patternfly.component.button.Button;
+import org.patternfly.core.Aria;
+import org.patternfly.handler.ActionHandler;
+
+import elemental2.dom.HTMLElement;
+
+import static org.jboss.elemento.EventType.click;
+import static org.patternfly.component.button.Button.button;
+import static org.patternfly.layout.PredefinedIcon.help;
+
+public class CodeEditorLink extends SubComponent<HTMLElement, CodeEditorLink> implements
+        ComponentReference<CodeEditor> {
+
+    // ------------------------------------------------------ factory methods
+
+    /**
+     * Adds a link. The button should have the modifiers {@link Button#link()}.
+     */
+    public static CodeEditorLink codeEditorLink(Button button) {
+        return new CodeEditorLink(button);
+    }
+
+    public static CodeEditorLink codeEditorViewShortcutsLink() {
+        return new CodeEditorLink(button().link().addIconAndText(help, "View shortcuts"))
+                .ariaLabel("View shortcuts");
+    }
+
+    // ------------------------------------------------------ instance
+
+    private final HTMLElement buttonElement;
+    private ActionHandler<CodeEditor> actionHandler;
+
+    CodeEditorLink(Button button) {
+        super(button.element());
+        buttonElement = element();
+    }
+
+    @Override
+    public void passComponent(CodeEditor codeEditor) {
+        if (actionHandler != null && buttonElement != null) {
+            buttonElement.addEventListener(click.name, e -> actionHandler.onAction(e, codeEditor));
+        }
+    }
+
+    @Override
+    public CodeEditorLink that() {
+        return this;
+    }
+
+    // ------------------------------------------------------ public API
+
+    public CodeEditorLink ariaLabel(String label) {
+        if (buttonElement != null) {
+            buttonElement.setAttribute(Aria.label, label);
+        }
+        return this;
+    }
+
+    // ------------------------------------------------------ events
+
+    public CodeEditorLink onAction(ActionHandler<CodeEditor> actionHandler) {
+        this.actionHandler = actionHandler;
+        return this;
+    }
+}